backend/requirements.txt
Flask==2.2.5
tensorflow==2.17.1
numpy==1.26.3
Pillow==10.2.0
